+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Build 1496 - 2026-09-22
+
+- automated build
+
 ## Build 1495 - 2026-09-22
 
 - automated build